Assigned name to the Instrument instead of id. Changed type of qty from int to BigDecimal. Added custom exceptions for more extensive control and covered most of the exception handling. Updated several log statements and fixed minor bugs. Added FatalFail Status enum for fatal errors when the whole operation should be aborted.
TODO:
* Take care of safe publishing while registering new services in the ExchangeBus, in case the current workaround does not suffice.

+++ src/main/java/server/List.java
@@ -10,7 +10,7 @@ public class List implements Message {
     private Instrument instrument;
     private Side side;
     private BigDecimal price;
-    private int qty;
+    private BigDecimal qty;
 
     private String session;
     private Instant date;
@@ -66,7 +66,7 @@ public class List implements Message {
         return clientId;
     }
 
-    public int getQty() {
+    public BigDecimal getQty() {
         return qty;
     }
